requirements.txt
flask>=2.2.0
Pillow>=9.0.0
PyPDF2>=2.10.0
requests>=2.28.0
pdf2image>=1.16.0
reportlab>=3.6.0
python-dotenv>=0.20.0
pdfplumber>=0.7.0
pdfminer.six>=20211012
pytesseract>=0.3.8
opencv-python-headless>=4.6.0
numpy>=1.22.0
flask-cors>=3.0.10 